requirements.txt
aiohttp==3.9.3
aioresponses==0.7.4
aiosqlite==0.19.0
discord.py==2.3.2
pydantic==1.10.2
pytest==7.4.0
python-dotenv==1.0.0
requests==2.31.0